promote the HuggingFace download stack out of lib
hfToken.js and hfDownload.js were never pure helpers: one reads settings to
resolve a token, the other spawns scripts/hf_download_repo.py and reads settings
to pick its venv. Both are orchestration, so this promotes them rather than
pushing their dependency down — the opposite call from the FableLoom/POST slice,
for the opposite reason.

  server/lib/hfToken.js    -> server/services/hfToken.js
  server/lib/hfDownload.js -> server/services/hfDownload.js

That left lib/sseDownload.js importing hfDownload, which would have inverted in
turn. It splits by what each export actually needs: openSseStream and
onClientDisconnect are pure response plumbing used by 11 routes with no HF
involvement and stay in lib, while startHfDownloadStream — which drives the
child process — moves to the new services/hfDownloadStream.js. Its 521-line
suite moves with it.

Re-points 21 static import sites, 10 vi.mock factories, and a set the AST
codemod could not see: a vi.doMock, a namespace import, and three dynamic
`await import()`s (two in ollamaManager.js). The full suite caught all five —
worth noting that a codemod over static ImportDeclarations alone leaves exactly
these behind.

ollamaManager.js keeps its lazy import of hfToken; that rationale is load cost,
not layering, and still holds. Comment reworded for the new path.

Refs #4901.
