+++ server/lib/index.js
@@ -122,11 +122,9 @@ export * from './contextBudget.js';
 export * from './cursor.js';
 export * from './grok.js';
 export * from './grokVideoClip.js';
-export * from './hfToken.js';
 export * from './heavyJobClaim.js';
 export * from './hfErrors.js';
 export * from './hfCache.js';
-export * from './hfDownload.js';
 export * from './icLoraWeights.js';
 export * from './sseDownload.js';
 export * from './sseHeaders.js';
